fix: use PR number as tag for Dependabot-trigger branch builds
It turns out we have to publish our base image(s) also when a build is triggered by Dependabot.
Our database-client images depend on that base-image to be available. But I don't want to publish
0.0.0 to avoid all kinds of race-conditions. So instead I will use the PR number of the PR opened
by Dependabot. This way we can a unique number of every PR and as this versioning scheme (I'm now
using `dependabot-<PR NUMBER>` as the tag) does not collide with SemVer, we don't run the risk of
overwriting images.

I also make sure the Dependabot-triggered build ignore pushes, that is, it is only triggered by a
`workflow_run` event. This reason is that builds triggered by a `push` event lack the permissions
to succeed anyway. To ignore these events for Dependabot I ignore all branches opened by
Dependabot for `push` event (for which the branch starts with `dependabot/...`). I also removed
the wrong conditional in `Get latest version`. This should not have been there in the first place.